refactor query functions into single file
