+++ backend/services/queries_test.go
@@ -1,0 +1,119 @@
+// Copyright 2022, Yahoo Inc.
+// Licensed under the terms of the MIT. See LICENSE file in project root for terms.
+
+package services_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/theparanoids/ashirt-server/backend/dtos"
+	"github.com/theparanoids/ashirt-server/backend/models"
+	"github.com/theparanoids/ashirt-server/backend/policy"
+	"github.com/theparanoids/ashirt-server/backend/services"
+)
+
+func TestCreateQuery(t *testing.T) {
+	db := initTest(t)
+	HarryPotterSeedData.ApplyTo(t, db)
+	ctx := fullContext(UserRon.ID, &policy.FullAccess{})
+
+	op := OpChamberOfSecrets
+	i := services.CreateQueryInput{
+		OperationSlug: op.Slug,
+		Name:          "Evidence By author",
+		Query:         "<query goes here>",
+		Type:          "findings",
+	}
+	createdQuery, err := services.CreateQuery(ctx, db, i)
+	require.NoError(t, err)
+	fullQuery := getQueryByID(t, db, createdQuery.ID)
+	require.Equal(t, i.Name, fullQuery.Name)
+	require.Equal(t, i.Type, fullQuery.Type)
+	require.Equal(t, i.Query, fullQuery.Query)
+}
+
+
+func TestDeleteQuery(t *testing.T) {
+	db := initTest(t)
+	HarryPotterSeedData.ApplyTo(t, db)
+	ctx := fullContext(UserRon.ID, &policy.FullAccess{})
+
+	i := services.DeleteQueryInput{
+		OperationSlug: OpChamberOfSecrets.Slug,
+		ID:            QuerySalazarsHier.ID,
+	}
+
+	getQueryCount := makeDBRowCounter(t, db, "queries", "id=?", i.ID)
+	require.Equal(t, int64(1), getQueryCount(), "Database should have item to delete")
+
+	err := services.DeleteQuery(ctx, db, i)
+	require.NoError(t, err)
+	require.Equal(t, int64(0), getQueryCount(), "Database should have deleted the item")
+}
+
+func TestListQueriesForOperation(t *testing.T) {
+	db := initTest(t)
+	HarryPotterSeedData.ApplyTo(t, db)
+	ctx := fullContext(UserRon.ID, &policy.FullAccess{})
+
+	masterOp := OpChamberOfSecrets
+	allQueries := getQueriesForOperationID(t, db, masterOp.ID)
+	require.NotEqual(t, len(allQueries), 0, "Some number of queries should exist")
+
+	foundQueries, err := services.ListQueriesForOperation(ctx, db, masterOp.Slug)
+	require.NoError(t, err)
+	require.Equal(t, len(foundQueries), len(allQueries))
+	validateQuerySets(t, foundQueries, allQueries, validateQuery)
+}
+
+func TestUpdateQuery(t *testing.T) {
+	db := initTest(t)
+	HarryPotterSeedData.ApplyTo(t, db)
+	ctx := fullContext(UserRon.ID, &policy.FullAccess{})
+
+	masterOp := OpChamberOfSecrets
+	masterQuery := QuerySalazarsHier
+	input := services.UpdateQueryInput{
+		OperationSlug: masterOp.Slug,
+		ID:            masterQuery.ID,
+		Name:          "New Name",
+		Query:         "New Query",
+	}
+
+	err := services.UpdateQuery(ctx, db, input)
+	require.NoError(t, err)
+
+	updatedQuery := getQueryByID(t, db, masterQuery.ID)
+
+	require.NoError(t, err)
+	require.Equal(t, input.Name, updatedQuery.Name)
+	require.Equal(t, input.Query, updatedQuery.Query)
+}
+
+
+type queryValidator func(*testing.T, models.Query, *dtos.Query)
+
+func validateQuery(t *testing.T, expected models.Query, actual *dtos.Query) {
+	require.Equal(t, expected.ID, actual.ID)
+	require.Equal(t, expected.Name, actual.Name)
+	require.Equal(t, expected.Query, actual.Query)
+	require.Equal(t, expected.Type, actual.Type)
+}
+
+func validateQuerySets(t *testing.T, dtoSet []*dtos.Query, dbSet []models.Query, validator queryValidator) {
+	var expected *models.Query = nil
+
+	for _, dtoItem := range dtoSet {
+		expected = nil
+		for _, dbItem := range dbSet {
+			if dbItem.ID == dtoItem.ID {
+				expected = &dbItem
+				break
+			}
+		}
+		require.NotNil(t, expected, "Result should have matching value")
+		validator(t, *expected, dtoItem)
+	}
+}
